[COMPLETE] updated all docs, ensured functionality, added CLI for customization
