+++ src/backend/etl/aggieExpertsAPI/formattedFeatures/grantsFeatures.json
@@ -1,0 +1,112 @@
+[
+  {
+    "id": "K320B61-117089",
+    "title": "All-to-all optical shuffle device based on arrayed waveguide grating routers",
+    "name": "All-to-all optical shuffle device based on arrayed waveguide grating routers § Completed • 2016 - 2018 • Yoo, S. J. Ben § National Security Agency • H98230-16-C-0820",
+    "funder": "National Security Agency",
+    "status": "Completed",
+    "startDate": "2016-10-01",
+    "endDate": "2018-12-31",
+    "grantRole": "Co  Principal  Investigator  Role",
+    "relatedExperts": [
+      {
+        "expertId": "83x5AQ8a",
+        "firstName": "Sung-Joo",
+        "middleName": "Ben",
+        "lastName": "Yoo",
+        "fullName": "Sung-Joo Ben Yoo",
+        "title": "Professor",
+        "organizationUnit": "Elec & Comp Engr",
+        "url": "expert/83x5AQ8a"
+      }
+    ]
+  },
+  {
+    "id": "K321127-1324",
+    "title": "NSF ECS-00-70444",
+    "name": "NSF ECS-00-70444 § Completed • 2000 - 2005 • Yoo, S. J. Ben § National Science Foundation • 0070444",
+    "funder": "National Science Foundation",
+    "status": "Completed",
+    "startDate": "2000-09-01",
+    "endDate": "2005-08-31",
+    "grantRole": "Grant  Role, Principal  Investigator  Role",
+    "relatedExperts": [
+      {
+        "expertId": "83x5AQ8a",
+        "firstName": "Sung-Joo",
+        "middleName": "Ben",
+        "lastName": "Yoo",
+        "fullName": "Sung-Joo Ben Yoo",
+        "title": "Professor",
+        "organizationUnit": "Elec & Comp Engr",
+        "url": "expert/83x5AQ8a"
+      }
+    ]
+  },
+  {
+    "id": "K321280-1477",
+    "title": "COLLABORATIVE RESEARCH:  NETS-NBD:  INTELLIGENT AND ADAPTIVE NETWORKING FOR THE NEXT GENERATION INTERNET",
+    "name": "COLLABORATIVE RESEARCH:  NETS-NBD:  INTELLIGENT AND ADAPTIVE NETWORKING FOR THE NEXT GENERATION INTERNET § Completed • 2005 - 2008 • Yoo, S. J. Ben § National Science Foundation • 0520333",
+    "funder": "National Science Foundation",
+    "status": "Completed",
+    "startDate": "2005-08-15",
+    "endDate": "2008-07-31",
+    "grantRole": "Co  Principal  Investigator  Role",
+    "relatedExperts": [
+      {
+        "expertId": "83x5AQ8a",
+        "firstName": "Sung-Joo",
+        "middleName": "Ben",
+        "lastName": "Yoo",
+        "fullName": "Sung-Joo Ben Yoo",
+        "title": "Professor",
+        "organizationUnit": "Elec & Comp Engr",
+        "url": "expert/83x5AQ8a"
+      }
+    ]
+  },
+  {
+    "id": "K321451-1648",
+    "title": "COLLABORATIVE RESEARCH:  NETS-NR:  HIGH-PERFORMANCE NEXT GENERATION NETWORKING TECHNOLOGY AND SYSTEMS FOR HETEROGENEOUS NETWORKS",
+    "name": "COLLABORATIVE RESEARCH:  NETS-NR:  HIGH-PERFORMANCE NEXT GENERATION NETWORKING TECHNOLOGY AND SYSTEMS FOR HETEROGENEOUS NETWORKS § Completed • 2004 - 2008 • Yoo, S. J. Ben § National Science Foundation • 0435529",
+    "funder": "National Science Foundation",
+    "status": "Completed",
+    "startDate": "2004-10-01",
+    "endDate": "2008-09-30",
+    "grantRole": "Grant  Role, Co  Principal  Investigator  Role",
+    "relatedExperts": [
+      {
+        "expertId": "83x5AQ8a",
+        "firstName": "Sung-Joo",
+        "middleName": "Ben",
+        "lastName": "Yoo",
+        "fullName": "Sung-Joo Ben Yoo",
+        "title": "Professor",
+        "organizationUnit": "Elec & Comp Engr",
+        "url": "expert/83x5AQ8a"
+      }
+    ]
+  },
+  {
+    "id": "K321661-1859",
+    "title": "UCB MCA: INTERACTIVE VISUALIZATION AND EXPLORATION OF MASSIVE MULTISOURCE DATA STREAMS IN COLLABORATIVE ENVIRONMENTS USING PARALLEL AND DISTRIBUTED COMPUTING",
+    "name": "UCB MCA: INTERACTIVE VISUALIZATION AND EXPLORATION OF MASSIVE MULTISOURCE DATA STREAMS IN COLLABORATIVE ENVIRONMENTS USING PARALLEL AND DISTRIBUTED COMPUTING § Completed • 2001 - 2006 • Hamann, Bernd § National Science Foundation • NSF-EIA-01-22599",
+    "funder": "National Science Foundation",
+    "status": "Completed",
+    "startDate": "2001-10-01",
+    "endDate": "2006-09-30",
+    "grantRole": "Co  Principal  Investigator  Role",
+    "relatedExperts": [
+      {
+        "expertId": "83x5AQ8a",
+        "firstName": "Sung-Joo",
+        "middleName": "Ben",
+        "lastName": "Yoo",
+        "fullName": "Sung-Joo Ben Yoo",
+        "title": "Professor",
+        "organizationUnit": "Elec & Comp Engr",
+        "url": "expert/83x5AQ8a"
+      }
+    ]
+  }
+]
